Refactor 'remote' middlewares
