+++ lib/app_profiler/viewer/middleware/firefox.rb
@@ -2,12 +2,12 @@
 
 require "app_profiler/yarn/command"
 require "app_profiler/yarn/with_firefox_profiler"
-require "app_profiler/viewer/firefox_remote_viewer/base_middleware"
+require "app_profiler/viewer/middleware/base"
 
 module AppProfiler
   module Viewer
     class RemoteViewer < BaseViewer
-      class FirefoxMiddleware < FirefoxBaseMiddleware
+      class FirefoxMiddleware < BaseMiddleware
         include Yarn::WithFirefoxProfile
 
         def initialize(app)
@@ -15,6 +15,18 @@ module AppProfiler
           @firefox_profiler = Rack::File.new(
             File.join(AppProfiler.root, "node_modules/firefox-profiler/dist")
           )
+        end
+
+        def call(env)
+          request = Rack::Request.new(env)
+          @app.call(env) if request.path_info.end_with?(".stackprof.json")
+          # Firefox profiler *really* doesn't like for /from-url/ to be at any other mount point
+          # so with this enabled, we take over both /app_profiler and /from-url in the app in development.
+          return from(env, Regexp.last_match(1))   if request.path_info =~ %r(\A/from-url(.*)\z)
+          return viewer(env, Regexp.last_match(1)) if request.path_info =~ %r(\A/app_profiler/firefox/viewer/(.*)\z)
+          return show(env, Regexp.last_match(1))   if request.path_info =~ %r(\A/app_profiler/firefox/(.*)\z)
+
+          super
         end
 
         protected
